cv/scripts/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --extra-index-url=https://pypi.org/simple --output-file=requirements.txt requirements.in
#
click==7.1.2              # via -r requirements.in
jinja2==2.11.3            # via -r requirements.in
markupsafe==1.1.1         # via jinja2
pydantic==1.7.3           # via -r requirements.in
pyyaml==5.4.1             # via -r requirements.in
